requirements.txt
gunicorn==20.1.0
pandas
httptools
uvloop
coverage
pytest
httpx
trio
pytest-asyncio
pytest-trio
uvicorn==0.13.4
fastapi==0.65.2
python-decouple==3.3
faunadb==4.1.0
coinbase==2.1.0
boto3==1.17.54
pystan==2.19.1.1
fbprophet==0.7.1
redis